mini-app-backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
aiogram==3.22.0
python-dotenv==1.1.1
